Tweaked Horizontal Movement.
Horizontal movement now directly sets the x component of the rigidbody's
velocity. The good: Snappy responsive controls that aren't too loose or
slippery and work well with thumbstick. The bad: Can't use built-in
physics to affect the x-component for more than one frame, as it will be
overwritten, not summed.
For mow this is fine, but it may need to be changed later.